requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
idna==2.9
kafka-python==2.0.1
psycopg2-binary==2.8.5
requests==2.23.0
urllib3==1.25.9
